Moved out components to resolve import errors with vs differences beteween dir names
